[Refactor] #4951 feeling-table.cpp/h をcmd-io/ からfloor/dungeon-feeling.cpp/h へ移した
