+++ src/cmd-io/cmd-dump.cpp
@@ -10,9 +10,9 @@
  */
 
 #include "cmd-io/cmd-dump.h"
-#include "cmd-io/feeling-table.h"
 #include "core/asking-player.h"
 #include "dungeon/quest.h"
+#include "floor/dungeon-feeling.h"
 #include "io-dump/dump-remover.h"
 #include "io-dump/dump-util.h"
 #include "io/files-util.h"
